# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TrendPoint {
  label: string;
  value: number;
}

export interface BreakdownPoint {
  label: string;
  value: number;
}

export interface DashboardSummary {
  todayAppointments: number;
  confirmed: number;
  pending: number;
  completed: number;
  pendingPayments: number;
  revenue: number;
  reportsPending: number;
  whatsappSent: number;
  appointmentTrend?: TrendPoint[];
  statusBreakdown?: BreakdownPoint[];
}

export interface Appointment {
  id: string;
  patientId: string;
  patientName: string;
  uhid: string;
  mobile: string;
  testName: string;
  lab: string;
  branch: string;
  date: string;
  time: string;
  /** @nullable */
  doctor?: string | null;
  status: string;
  paymentStatus: string;
  reportStatus?: string;
  amount: number;
  source: string;
  createdBy: string;
  /** @nullable */
  notes?: string | null;
  updatedAt?: string;
}

export interface AppointmentPage {
  items: Appointment[];
  page: number;
  pageSize: number;
  total: number;
}

export interface AppointmentInput {
  patientId: string;
  testId: string;
  date: string;
  time: string;
  slotId: string;
  doctor?: string;
  source: string;
  paymentStatus: string;
  notes?: string;
}

export interface StatusUpdate {
  status: string;
  reason?: string;
}

export interface PaymentUpdate {
  paymentStatus: string;
  paymentMethod?: string;
}

export interface Patient {
  id: string;
  uhid: string;
  name: string;
  mobile: string;
  whatsapp?: string;
  /** @nullable */
  email?: string | null;
  age: number;
  gender: string;
  lastVisit: string;
  totalVisits: number;
}

export interface PatientInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 8 */
  mobile: string;
  whatsapp?: string;
  email?: string;
  /** @minimum 0 */
  age: number;
  gender: string;
}

export interface Test {
  id: string;
  name: string;
  shortName: string;
  department: string;
  duration: number;
  price: number;
  active: boolean;
}

export interface Slot {
  id: string;
  date: string;
  startTime: string;
  endTime: string;
  capacity: number;
  booked: number;
  status: string;
  testName: string;
  lab: string;
}

export interface Activity {
  id: string;
  title: string;
  detail: string;
  time: string;
  type: string;
}

export type PageParameter = number;

export type PageSizeParameter = number;

export type ListAppointmentsParams = {
/**
 * @minimum 1
 */
page?: PageParameter;
/**
 * @minimum 1
 * @maximum 50
 */
pageSize?: PageSizeParameter;
date?: string;
status?: string;
search?: string;
};

export type ListPatientsParams = {
search?: string;
};

export type ListSlotsParams = {
date?: string;
testId?: string;
};