# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Operational dashboard
  - name: appointments
    description: Appointment operations
  - name: patients
    description: Patient master
  - name: tests
    description: Test master
  - name: slots
    description: Appointment slot availability
  - name: activity
    description: Recent operational activity
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: date
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated appointment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentPage"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "409":
          description: Slot unavailable
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Appointment not found
  /appointments/{id}/status:
    patch:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/payment:
    patch:
      operationId: updateAppointmentPayment
      tags: [appointments]
      summary: Update appointment payment
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Patient list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /tests:
    get:
      operationId: listTests
      tags: [tests]
      summary: List active diagnostic tests
      responses:
        "200":
          description: Test list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"
  /slots:
    get:
      operationId: listSlots
      tags: [slots]
      summary: List available slots
      parameters:
        - name: date
          in: query
          schema: { type: string }
        - name: testId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Slot list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Slot"
  /activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: List recent activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    Page:
      name: page
      in: query
      schema: { type: number, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: number, minimum: 1, maximum: 50, default: 50 }
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DashboardSummary:
      type: object
      required: [todayAppointments, confirmed, pending, completed, pendingPayments, revenue, reportsPending, whatsappSent]
      properties:
        todayAppointments: { type: number }
        confirmed: { type: number }
        pending: { type: number }
        completed: { type: number }
        pendingPayments: { type: number }
        revenue: { type: number }
        reportsPending: { type: number }
        whatsappSent: { type: number }
        appointmentTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        statusBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/BreakdownPoint"
    TrendPoint:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: number }
    BreakdownPoint:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: number }
    Appointment:
      type: object
      required: [id, patientName, patientId, uhid, mobile, testName, lab, branch, date, time, status, paymentStatus, amount, source, createdBy]
      properties:
        id: { type: string }
        patientId: { type: string }
        patientName: { type: string }
        uhid: { type: string }
        mobile: { type: string }
        testName: { type: string }
        lab: { type: string }
        branch: { type: string }
        date: { type: string }
        time: { type: string }
        doctor: { type: ["string", "null"] }
        status: { type: string }
        paymentStatus: { type: string }
        reportStatus: { type: string }
        amount: { type: number }
        source: { type: string }
        createdBy: { type: string }
        notes: { type: ["string", "null"] }
        updatedAt: { type: string }
    AppointmentPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        page: { type: number }
        pageSize: { type: number }
        total: { type: number }
    AppointmentInput:
      type: object
      required: [patientId, testId, date, time, slotId, source, paymentStatus]
      properties:
        patientId: { type: string }
        testId: { type: string }
        date: { type: string }
        time: { type: string }
        slotId: { type: string }
        doctor: { type: string }
        source: { type: string }
        paymentStatus: { type: string }
        notes: { type: string }
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
        reason: { type: string }
    PaymentUpdate:
      type: object
      required: [paymentStatus]
      properties:
        paymentStatus: { type: string }
        paymentMethod: { type: string }
    Patient:
      type: object
      required: [id, uhid, name, mobile, age, gender, lastVisit, totalVisits]
      properties:
        id: { type: string }
        uhid: { type: string }
        name: { type: string }
        mobile: { type: string }
        whatsapp: { type: string }
        email: { type: ["string", "null"] }
        age: { type: number }
        gender: { type: string }
        lastVisit: { type: string }
        totalVisits: { type: number }
    PatientInput:
      type: object
      required: [name, mobile, age, gender]
      properties:
        name: { type: string, minLength: 2 }
        mobile: { type: string, minLength: 8 }
        whatsapp: { type: string }
        email: { type: string }
        age: { type: number, minimum: 0 }
        gender: { type: string }
    Test:
      type: object
      required: [id, name, shortName, department, duration, price, active]
      properties:
        id: { type: string }
        name: { type: string }
        shortName: { type: string }
        department: { type: string }
        duration: { type: number }
        price: { type: number }
        active: { type: boolean }
    Slot:
      type: object
      required: [id, date, startTime, endTime, capacity, booked, status, testName, lab]
      properties:
        id: { type: string }
        date: { type: string }
        startTime: { type: string }
        endTime: { type: string }
        capacity: { type: number }
        booked: { type: number }
        status: { type: string }
        testName: { type: string }
        lab: { type: string }
    Activity:
      type: object
      required: [id, title, detail, time, type]
      properties:
        id: { type: string }
        title: { type: string }
        detail: { type: string }
        time: { type: string }
        type: { type: string }