# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Appointment {
  id: string;
  patientId: string;
  patientName: string;
  uhid: string;
  mobile: string;
  testName: string;
  lab: string;
  branch: string;
  date: string;
  time: string;
  /** @nullable */
  doctor?: string | null;
  status: string;
  paymentStatus: string;
  reportStatus?: string;
  amount: number;
  source: string;
  createdBy: string;
  /** @nullable */
  notes?: string | null;
  updatedAt?: string;
}