# lib/api-zod/src/generated/types/appointmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentInput {
  patientId: string;
  testId: string;
  date: string;
  time: string;
  slotId: string;
  doctor?: string;
  source: string;
  paymentStatus: string;
  notes?: string;
}