# lib/api-zod/src/generated/types/appointmentPage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';

export interface AppointmentPage {
  items: Appointment[];
  page: number;
  pageSize: number;
  total: number;
}