# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BreakdownPoint } from './breakdownPoint';
import type { TrendPoint } from './trendPoint';

export interface DashboardSummary {
  todayAppointments: number;
  confirmed: number;
  pending: number;
  completed: number;
  pendingPayments: number;
  revenue: number;
  reportsPending: number;
  whatsappSent: number;
  appointmentTrend?: TrendPoint[];
  statusBreakdown?: BreakdownPoint[];
}