# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: string;
  uhid: string;
  name: string;
  mobile: string;
  whatsapp?: string;
  /** @nullable */
  email?: string | null;
  age: number;
  gender: string;
  lastVisit: string;
  totalVisits: number;
}