# lib/api-zod/src/generated/types/patientInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PatientInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 8 */
  mobile: string;
  whatsapp?: string;
  email?: string;
  /** @minimum 0 */
  age: number;
  gender: string;
}