# lib/api-zod/src/generated/types/paymentUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentUpdate {
  paymentStatus: string;
  paymentMethod?: string;
}