# lib/api-zod/src/generated/types/test.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Test {
  id: string;
  name: string;
  shortName: string;
  department: string;
  duration: number;
  price: number;
  active: boolean;
}